rename package name from screen to pages.
